refactor(cosmic-swingset): Move makeChainStorageNode into a vat
